refactor(admin): extract Settings > Social links into standalone component
Move the Settings > Social links panel (contact phone/email plus the
Instagram/Facebook page lists with add/remove and thumbnail fetch) out of
the monolithic AdminComponent into a new standalone OnPush component
`app-admin-social-links` under pages/admin/settings/. The component owns
socialForm/socialMessage/socialError/socialThumb* state and the
load/save/add/remove/fetch-thumbnail/parse/sanitize logic; the shared CMS
content-version bookkeeping stays on the parent and is threaded in via the
existing boundRememberContentVersion/withExpectedVersion/handleContentConflict/
forgetContentVersion callback inputs. The parent template now renders
`<app-admin-social-links>` and no longer self-calls loadSocial (the child
loads on init).

Behavioural coverage moves with the code into
admin-social-links.component.spec.ts; the corresponding social specs are
dropped from admin.component(.branchfill/.coverage).spec.ts.
